Refactor: libstonithd: separate RHCS-specific code into own source file
improves readability and maintainability
